requirements.txt
behave
selenium
nose
webdriver-manager
docx
python-docx
PyYAML
